✨ feat:  Navigates from PIN to Home (#4)
* ✨ feat: Add new file app/constants/screens.ts

This commit adds a new file `app/constants/screens.ts` with a mode of 100644 and an index of 0000000.

Possible options:
- chore: Update README.md
- fix: Correct typos in app/constants/screens.ts
- test: Write unit tests for app/constants/screens.ts
- refactor: Simplify code in app/constants/screens.ts
- docs: Add JSDoc comments to app/constants/screens.ts

* ✨ feat: Add constants for Home, Launch, Settings, and Main screens

This commit adds constants for the Home, Launch, Settings, and Main screens in the app/constants/screens.ts file. The new constants are exported and can be used throughout the application to reference these screens consistently.

* ✨ feat: Add new screens and constants to navigation

Add HOME, SETTINGS constants for screen names and update the respective Tab.Screen components with options to hide their headers.
Rename the Stack.Screen component from "HomeRoute" to "MAIN" and add options to hide its header.
Rename the Stack.Screen component from "Launch" to "LAUNCH".
Update the isSignIn function's todo comment to indicate that it will read from the database.

* ✨ feat: Add a nested navigator best practice group link to the pin screen

Add a comment containing a link to the best practices for nesting navigators in React Navigation to the pin screen.

* ✨ feat: Add PIN input screen to login flow

This commit introduces a new PIN input screen to the login flow. It replaces the previous code input screen and now allows users to enter a 6-digit PIN. The PIN is stored locally in the database once entered, and if the length of the PIN is greater than 5, the stacks auto-switch to the Home screen. Additionally, some code refactoring was done to remove unnecessary code and add styles to the components.