✨ feat: Add functionality to check if user is signed in and store PIN in the database
This commit adds a function to check if the user is signed in by querying the database, as well as updating the `isSignIn` function to return true if the user is signed in. Additionally, it includes changes to the `Launch` component that create a database to store the PIN entered by the user and ensures that the stacks auto-switch to the Home screen after the PIN is stored.