✨ feat: refactor navigation and remove unused code
This commit refactors the navigation code, removing unused imports and variables. The `createNativeStackNavigator` and `Launch` components were removed, as they were not being used. The `ALTURA_PURPLE` constant was also removed, as it was no longer being used. Additionally, the `HomeRoute` component was exported for reuse in other parts of the application. The `isSignIn` function was also removed, as it is not yet implemented.