✨ feat: Add user subscription and creation functionality
This commit adds the ability to subscribe to changes in the `User` table of the app's database using `subscribeUser()`. It also includes the option to create a new user with a PIN and an optional name using `createUser()`. Additionally, it imports and uses `react-native-crypto-js` for cryptographic functionality.

The `Launch` screen now triggers user creation when a 6-digit PIN is entered. The created user is stored in the local database, and the stacks auto-switch to the Home screen upon successful creation.

A comment has been added to provide information on how to view the contents of the database.