Feat database (#5)
* ✨ feat: Installation of watermelondb

* ✨ feat: Add new features and improvements

Added ideas for new features and improvements to the app, including the ability to view balances in other currencies, refreshing balance on home screen, implementing analytics to track usage, using CodePush for faster hotfixes, implementing Sentry for error reporting, and storing the database at a more secure location. Also added new files and models related to account and user management.

* ✨ feat: Add AccountModel and UserModel types

This commit adds AccountModel and UserModel types to `types/models/account.ts` and `types/models/user.ts` respectively. These types define the table name, fields, and methods for each model in WatermelonDB.

* ✨ feat: Add balance and crypto type fields to `Account` model, and add `setBalance` and `setCryptoType` writer methods, and add name and pin fields to `User` model, and add `setName` and `setPin` writer methods

* ✨ feat: Add database manager, migration, and schema

This commit adds the database manager, migration, and schema for the app's database. The database manager creates a Watermelon database using SQLiteAdapter with models for User and Account. The migration file is currently empty, but will be used to define migrations later. Finally, the schema defines tables and columns for User and Account models.

* ✨ feat: Add experimental decorators option to tsconfig.json and define two constants in table.ts

The commit adds the experimental decorators option to tsconfig.json and defines two new constants (WDB_TBL_USER and WDB_TBL_ACCOUNT) in table.ts file.

* ✨ feat: Add functionality to check if user is signed in and store PIN in the database

This commit adds a function to check if the user is signed in by querying the database, as well as updating the `isSignIn` function to return true if the user is signed in. Additionally, it includes changes to the `Launch` component that create a database to store the PIN entered by the user and ensures that the stacks auto-switch to the Home screen after the PIN is stored.

* ✨ feat: add queryUser function for fetching user from database

This commit adds a new file `user.ts` in the `app/queries` directory, which exports an asynchronous function named `queryUser`. The function fetches a user model from the `WDB_TBL_USER` table using the `fetch()` method of the `get` query builder. If successful, the function returns the first user found, otherwise, it returns null. Any errors that occur during the query execution will be logged to the console.

* ✨ feat: Add new features for improved security and functionality

This commit adds various features to the app, including implementing CodePush and Sentry, storing the database at a more secure location, and allowing for multiple user accounts. The ability to emulate multiple user accounts is achieved by having multiple user records with different PINs, and the app tracks down the user record with the latest time stamp.

* ✨ feat: refactor navigation and remove unused code

This commit refactors the navigation code, removing unused imports and variables. The `createNativeStackNavigator` and `Launch` components were removed, as they were not being used. The `ALTURA_PURPLE` constant was also removed, as it was no longer being used. Additionally, the `HomeRoute` component was exported for reuse in other parts of the application. The `isSignIn` function was also removed, as it is not yet implemented.

* ✨ feat: import App component and remove unused code

The commit introduces a feature by importing the `App` component in the `index.js` file of a React Native project. Additionally, it removes unused code related to formatting.

* ✨ feat: add init() method to DatabaseManager

This commit adds an `init()` method to the `DatabaseManager` class which initializes the database by creating a Watermelon database from the provided adapter and model classes. If successful, the method assigns the resulting database object to the `database` property of the class.

* ✨ feat: Add navigation and authentication logic to App component

This commit introduces a NavigationContainer and createNativeStackNavigator from @react-navigation/native to the App component, allowing for navigation between screens. It also adds authentication logic to determine whether the user is logged in or not, and conditionally renders either the HomeRoute or Launch screen based on that. The useEffect hook is used to asynchronously query the user data and setIsLoggedIn and setIsQuerying are used to keep track of the query's progress. Finally, a loader is displayed while waiting for the query to complete.

* ✨ feat: Use TypeScript non-null assertion operator in `Account` and `User` models

This commit modifies the `Account` and `User` models to use the non-null assertion operator in TypeScript for the `balance` and `cryptoType` properties in the `Account` model, and the `name` and `pin` properties in the `User` model. This change allows for better type checking and eliminates unnecessary null checks.

* build: Update Babel configuration and add 'babel-plugin-transform-remove-console' to dependencies.

* chore: git hook

* ✨ feat: Add user subscription and creation functionality

This commit adds the ability to subscribe to changes in the `User` table of the app's database using `subscribeUser()`. It also includes the option to create a new user with a PIN and an optional name using `createUser()`. Additionally, it imports and uses `react-native-crypto-js` for cryptographic functionality.

The `Launch` screen now triggers user creation when a 6-digit PIN is entered. The created user is stored in the local database, and the stacks auto-switch to the Home screen upon successful creation.

A comment has been added to provide information on how to view the contents of the database.

* Update user.ts